+++ .idea/workspace.xml
@@ -34,8 +34,11 @@
   </component>
   <component name="RecentsManager">
     <key name="MoveFile.RECENT_KEYS">
-      <recent name="C:\Users\yukie\WebstormProjects\CPSC_455_Project\frontend\pfiniture\src\components" />
-      <recent name="C:\Users\yukie\WebstormProjects\CPSC_455_Project\frontend\pfiniture\src\pages" />
+      <recent name="C:\Users\yukie\WebstormProjects\CPSC_455_Project\frontend\pfiniture\src\components\registration" />
+      <recent name="C:\Users\yukie\WebstormProjects\CPSC_455_Project\frontend\pfiniture\src\components\about" />
+      <recent name="C:\Users\yukie\WebstormProjects\CPSC_455_Project\frontend\pfiniture\src\components\login" />
+      <recent name="C:\Users\yukie\WebstormProjects\CPSC_455_Project\frontend\pfiniture\src\components\landing" />
+      <recent name="C:\Users\yukie\WebstormProjects\CPSC_455_Project\frontend\pfiniture\src\components\my-account" />
     </key>
   </component>
   <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
