+++ frontend/pfiniture/src/App.js
@@ -1,15 +1,14 @@
-
 import logo from './logo.svg'
 import './App.css'
 import 'bootstrap/dist/css/bootstrap.min.css'
-import LogInForm from './components/LogInForm'
-import RegistrationForm from './components/RegistrationForm'
-import About from './components/About'
+import LogInForm from './components/login/LogInForm'
+import RegistrationForm from './components/registration/RegistrationForm'
+import About from './components/about/About'
 import { Nav } from 'react-bootstrap'
 import React from 'react'
 import Landing from './pages/Landing'
-import Dashboard from "./components/Dashboard"
-import Navbar from "./Navbar";
+import Dashboard from "./components/my-account/Dashboard"
+import Navbar from "./components/Navbar";
 
 function App() {
   return (
