+++ frontend/pfiniture/src/components/landing/DisplayAllFurniture.css
