+++ frontend/pfiniture/src/components/landing/DisplayAllFurniture.js
