+++ frontend/pfiniture/src/components/landing/DisplayIndividualFurniture.js
