+++ frontend/pfiniture/src/components/landing/TemporaryFurniture.js
